Inline MarkdownPage into Page
