+++ beta/src/components/Layout/Page.tsx
@@ -10,19 +10,32 @@ import {useActiveSection} from 'hooks/useActiveSection';
 import {Footer} from './Footer';
 import {Toc} from './Toc';
 import SocialBanner from '../SocialBanner';
+import {DocsPageFooter} from 'components/DocsFooter';
+import {Seo} from 'components/Seo';
+import PageHeading from 'components/PageHeading';
+import {useRouteMeta} from './useRouteMeta';
+import {TocContext} from '../MDX/TocContext';
 import sidebarLearn from '../../sidebarLearn.json';
 import sidebarReference from '../../sidebarReference.json';
 import type {TocItem} from 'components/MDX/TocContext';
 import type {RouteItem} from 'components/Layout/useRouteMeta';
 
+import(/* webpackPrefetch: true */ '../MDX/CodeBlock/CodeBlock');
+
 interface PageProps {
   children: React.ReactNode;
   toc: Array<TocItem>;
   routeTree: RouteItem;
+  meta: {title?: string; description?: string};
 }
 
-export function Page({children, toc, routeTree}: PageProps) {
+export function Page({children, toc, routeTree, meta}: PageProps) {
   const {asPath} = useRouter();
+  const {route, nextRoute, prevRoute} = useRouteMeta(routeTree);
+  const section = useActiveSection();
+  const title = meta.title || route?.title || '';
+  const description = meta.description || route?.description || '';
+  const isHomePage = section === 'home';
   return (
     <>
       <SocialBanner />
@@ -35,7 +48,29 @@ export function Page({children, toc, routeTree}: PageProps) {
           <main className="min-w-0">
             <div className="lg:hidden h-16 mb-2" />
             <article className="break-words" key={asPath}>
-              {children}
+              <div className="pl-0">
+                <Seo title={title} />
+                {!isHomePage && (
+                  <PageHeading
+                    title={title}
+                    description={description}
+                    tags={route?.tags}
+                    routeTree={routeTree}
+                  />
+                )}
+                <div className="px-5 sm:px-12">
+                  <div className="max-w-7xl mx-auto">
+                    <TocContext.Provider value={toc}>
+                      {children}
+                    </TocContext.Provider>
+                  </div>
+                  <DocsPageFooter
+                    route={route}
+                    nextRoute={nextRoute}
+                    prevRoute={prevRoute}
+                  />
+                </div>
+              </div>
             </article>
             <Footer />
           </main>
